Small changes to comply with *most* of the ICC warnings. * the 'unsigned int's in the geometry have been changed to 'int' * virtual functions in template classes in autosun are now managed correctly * included headers here and there * functions became static where possible * cleaned some single-precision mess in the g5QMR inverter * removed unused variables
git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/Branches/HiRep-HMC@429 37a69121-b030-0410-a54a-fdca83ffee47

+++ Converter/archive_eolexi.c
@@ -132,7 +132,7 @@ void write_gauge_field_eolexi_BE(char filename[])
   if(PID==0) {
     error((fp=fopen(filename,"wb"))==NULL,1,"write_gauge_field_eolexi",
         "Failed to open file for writing");
-    error(fwrite_BE_double((double*)eolexi_field,4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),fp)!=4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),
+    error(fwrite_BE_double((double*)eolexi_field,4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),fp)!=4*GLB_T*GLB_X*GLB_Y*GLB_Z*(int)sizeof(suNg)/(int)sizeof(double),
         1,"write_gauge_field_eolexi",
         "Failed to write gauge field to file");
     fclose(fp);
@@ -157,7 +157,7 @@ void read_gauge_field_eolexi_BE(char filename[])
   if(PID==0) {
     error((fp=fopen(filename,"rb"))==NULL,1,"read_gauge_field_eolexi",
         "Failed to open file for reading");
-    error(fread_BE_double((double*)eolexi_field,4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),fp)!=4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),
+    error(fread_BE_double((double*)eolexi_field,4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),fp)!=4*GLB_T*GLB_X*GLB_Y*GLB_Z*(int)sizeof(suNg)/(int)sizeof(double),
         1,"read_gauge_field_eolexi",
         "Failed to read gauge field from file");
   }
@@ -307,7 +307,7 @@ void write_gauge_field_eolexi_LE(char filename[])
   if(PID==0) {
     error((fp=fopen(filename,"wb"))==NULL,1,"write_gauge_field_eolexi",
         "Failed to open file for writing");
-    error(fwrite_LE_double((double*)eolexi_field,4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),fp)!=4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),
+    error(fwrite_LE_double((double*)eolexi_field,4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),fp)!=4*GLB_T*GLB_X*GLB_Y*GLB_Z*(int)sizeof(suNg)/(int)sizeof(double),
         1,"write_gauge_field_eolexi",
         "Failed to write gauge field to file");
     fclose(fp);
@@ -332,7 +332,7 @@ void read_gauge_field_eolexi_LE(char filename[])
   if(PID==0) {
     error((fp=fopen(filename,"rb"))==NULL,1,"read_gauge_field_eolexi",
         "Failed to open file for reading");
-    error(fread_LE_double((double*)eolexi_field,4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),fp)!=4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),
+    error(fread_LE_double((double*)eolexi_field,4*GLB_T*GLB_X*GLB_Y*GLB_Z*sizeof(suNg)/sizeof(double),fp)!=4*GLB_T*GLB_X*GLB_Y*GLB_Z*(int)sizeof(suNg)/(int)sizeof(double),
         1,"read_gauge_field_eolexi",
         "Failed to read gauge field from file");
   }
