+++ Update/random_spinor_field.c
@@ -13,7 +13,7 @@
 
 void gaussian_spinor_field(spinor_field *s) {
 	const double c1=1./sqrt(2.);
-	unsigned int i;
+	int i;
 	geometry_descriptor *type = s->type;
 	for(i=0;i<type->local_master_pieces;i++)
  	  gauss((double*)(s->ptr+type->master_start[i]),(type->master_end[i]-type->master_start[i]+1)*sizeof(suNf_spinor)/sizeof(double));
@@ -22,8 +22,8 @@ void gaussian_spinor_field(spinor_field *s) {
 }
 
 void gaussian_spinor_field_flt(spinor_field_flt *s) {
-	const float c1=1./sqrt(2.);
-	unsigned int i;
+	const float c1=(float)(1./sqrt(2.));
+	int i;
 	geometry_descriptor *type = s->type;
 	for(i=0;i<type->local_master_pieces;i++)
  	  gauss_flt((float*)(s->ptr+type->master_start[i]),(type->master_end[i]-type->master_start[i]+1)*sizeof(suNf_spinor_flt)/sizeof(float));
@@ -33,7 +33,7 @@ void gaussian_spinor_field_flt(spinor_field_flt *s) {
 
 
 void z2_spinor_field(spinor_field *s) {
-	unsigned int i;
+	int i;
 	geometry_descriptor *type = s->type;
 	for(i=0;i<type->local_master_pieces;i++)
  	  ranz2((double*)(s->ptr+type->master_start[i]),(type->master_end[i]-type->master_start[i]+1)*sizeof(suNf_spinor)/sizeof(double));
